Support passing Assertions directly to `JobTest` `output` (#1993)
* Support passing Assertions directly to `JobTest` `output`

The instructions and examples for `JobTest` `output` assertions warn if
run with `-Ywarn-value-discard`, because the compiler sees the `Unit` as
being discarded.  Most of the tests I can see fix this by returning `()`
_after_ the assertion, but this shouldn't be necessary.  This commit
changes the `assertion` to be `SCollection[T] => A`, where `A` is
discarded.  The `assertion` is still converted using
`.asInstanceOf[SCollection[_] => Unit]` (so there's no actual behavior
change), but now passing `Assertion`s will work without a hitch.

Also see: [relevant discussion in
Gitter](https://gitter.im/spotify/scio?at=5d0bf10e4b0b7b477b2c14f2)

* Use `_` instead of discarded type parameter `A`

This way, the signature does not change.

* Use `Any` instead of `_`

The hope is that this will satisfy the "missing parameter type" errors
that appeared with `_`.